reorganization of config and params
