+++ crate/server/src/config/clap/bootstrap_server_config.rs
@@ -19,7 +19,8 @@ pub struct BootstrapServerConfig {
     )]
     pub bootstrap_server_common_name: String,
 
-    /// The bootstrap server may be started on a specific port
+    /// The bootstrap server may be started on a specific port,
+    /// The hostname will be that configured in --hostname
     #[clap(long, env("KMS_BOOTSTRAP_SERVER_PORT"), default_value("9998"))]
     pub bootstrap_server_port: usize,
 }
