+++ crate/server/src/config/params/enclave_params.rs
@@ -1,0 +1,9 @@
+use std::path::PathBuf;
+
+#[derive(Clone, Debug)]
+pub struct EnclaveParams {
+    // contains the path to the manifest
+    pub manifest_path: PathBuf,
+    // contains the path to the signer public key
+    pub public_key_path: PathBuf,
+}
