+++ crate/server/src/config/params/mod.rs
@@ -1,0 +1,11 @@
+mod db_params;
+mod enclave_params;
+mod server_params;
+
+pub use db_params::DbParams;
+pub use enclave_params::EnclaveParams;
+pub use server_params::ServerParams;
+
+use super::clap::BootstrapServerConfig;
+
+pub type BootstrapServerParams = BootstrapServerConfig;
