+++ crate/server/src/tests/curve_25519_tests.rs
@@ -14,7 +14,7 @@ use cosmian_kms_utils::crypto::curve_25519::{
 };
 
 use crate::{
-    config::ServerConfig, error::KmsError, result::KResult, tests::test_utils::https_clap_config,
+    config::ServerParams, error::KmsError, result::KResult, tests::test_utils::https_clap_config,
     KMSServer,
 };
 
@@ -22,7 +22,7 @@ use crate::{
 async fn test_curve_25519_key_pair() -> KResult<()> {
     let clap_config = https_clap_config();
 
-    let kms = Arc::new(KMSServer::instantiate(ServerConfig::try_from(&clap_config).await?).await?);
+    let kms = Arc::new(KMSServer::instantiate(ServerParams::try_from(&clap_config).await?).await?);
     let owner = "eyJhbGciOiJSUzI1Ni";
 
     // request key pair creation
