+++ crate/server/src/tests/kmip_server_tests.rs
@@ -18,7 +18,7 @@ use tracing::trace;
 use uuid::Uuid;
 
 use crate::{
-    config::ServerConfig, error::KmsError, log_utils::log_init, result::KResult,
+    config::ServerParams, error::KmsError, log_utils::log_init, result::KResult,
     tests::test_utils::https_clap_config, KMSServer,
 };
 
@@ -26,7 +26,7 @@ use crate::{
 async fn test_curve_25519_key_pair() -> KResult<()> {
     let clap_config = https_clap_config();
 
-    let kms = Arc::new(KMSServer::instantiate(ServerConfig::try_from(&clap_config).await?).await?);
+    let kms = Arc::new(KMSServer::instantiate(ServerParams::try_from(&clap_config).await?).await?);
     let owner = "eyJhbGciOiJSUzI1Ni";
 
     // request key pair creation
@@ -163,7 +163,7 @@ async fn test_import_wrapped_symmetric_key() -> KResult<()> {
 
     let clap_config = https_clap_config();
 
-    let kms = Arc::new(KMSServer::instantiate(ServerConfig::try_from(&clap_config).await?).await?);
+    let kms = Arc::new(KMSServer::instantiate(ServerParams::try_from(&clap_config).await?).await?);
     let owner = "eyJhbGciOiJSUzI1Ni";
 
     let wrapped_symmetric_key = [0_u8; 32];
@@ -219,7 +219,7 @@ async fn test_database_user_tenant() -> KResult<()> {
 
     let clap_config = https_clap_config();
 
-    let kms = Arc::new(KMSServer::instantiate(ServerConfig::try_from(&clap_config).await?).await?);
+    let kms = Arc::new(KMSServer::instantiate(ServerParams::try_from(&clap_config).await?).await?);
     let owner = "eyJhbGciOiJSUzI1Ni";
 
     // request key pair creation
